Implemented (but not tested) the rest of reservations view and fixed bug with an important distinction in long int between c++ and c#

+++ UniversityReservationSystem.Interface/Helpers/DateTimeHelpers.cs
@@ -4,12 +4,12 @@ namespace UniversityReservationSystem.Interface.Helpers
 {
     public static class DateTimeHelpers
     {
-        public static int ToUnixTimestamp(this DateTime dateTime)
+        public static long ToUnixTimestamp(this DateTime dateTime)
         {
             return  (int)((dateTime - new DateTime(1970, 1, 1).ToLocalTime()).TotalSeconds);
         }
 
-        public static DateTime ToDateTime(this long timeStamp)
+        public static DateTime ToDateTime(this int timeStamp)
         {
             var unixEpochDateTime = new DateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeKind.Utc);
             return unixEpochDateTime.AddSeconds(timeStamp).ToLocalTime();
